+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+## Removed
+- Check for chunk header in `chunk_size_and_padding` function
+
 ## [0.2.4] - 2022-03-28
 ## Changed
 - pack\_pages function now return an empty byte when given an empty list instead of throwing an exception.
@@ -42,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug in extractusm that causes it to fail when output directory doesn't exist.
 - Fixed bug where program fails when directory exists.
 
+[Unreleased]: https://github.com/donmai-me/WannaCRI/compare/0.2.4...HEAD
 [0.2.4]: https://github.com/donmai-me/WannaCRI/compare/0.2.3...0.2.4
 [0.2.3]: https://github.com/donmai-me/WannaCRI/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/donmai-me/WannaCRI/compare/0.2.1...0.2.2